requirements.txt
flask==3.0.3
gunicorn==22.0.0
scikit-learn==1.5.2
pandas==2.2.3
numpy==1.26.4
pillow==10.4.0
joblib==1.4.2
